Moved test case to metadata-list.test
