+++ src/components/metadata/index.js
@@ -114,7 +114,6 @@ const MetaData = ({
                   commas={false}
                   inline={false}
                   value={metadata.parameters}
-                  limit={10}
                 />
                 <MetaDataRow
                   label="Inputs:"
